damc: fix timemeasure with TIM2 at 216Mhz
APB1 must be at 216Mhz for TIM5 clock output for DAMC HAT.

+++ damc/damc_simple_lib/TimeMeasure.h
@@ -31,8 +31,8 @@ public:
 
 private:
 	static uint32_t clock_per_us;
-	uint64_t time_sum;
-	uint64_t time_sum_per_loop;
+	uint32_t time_sum;
+	uint32_t time_sum_per_loop;
 	uint32_t time_max;
 	uint32_t begin_time;
 	bool isMeasuring;
